requirements.txt
mojimoji
jaconv
numpy
tqdm
pillow
